+++ src/components/menu/pc/handler/index.tsx
@@ -14,8 +14,10 @@ import Notice from "../../notice";
 import Github from "./github";
 
 export const languages: DropdownOption[] = [
-  { label: "🇨🇳 简体中文", value: "zh-CN" },
   { label: "🇺🇸 English", value: "en" },
+  { label: "🇨🇳 简体中文", value: "zh-CN" },
+  { label: "🇭🇰 繁体中文", value: "zh-HK" },
+  // { label: "🇯🇵 日本語", value: "ja" },
 ];
 
 export default function Handler() {