+++ src/middleware.ts
@@ -1,7 +1,8 @@
 import createMiddleware from "next-intl/middleware";
 
 export default createMiddleware({
-  locales: ["en", "zh-CN"],
+  locales: ["en", "zh-CN", "zh-HK"],
+  // locales: ["en", "zh-CN", "zh-HK", "ja"],
   defaultLocale: "en",
 });
 
@@ -19,7 +20,7 @@ export const config = {
 
 //   // Step 2: Create and call the next-intl middleware
 //   const handleI18nRouting = createMiddleware({
-//     locales: ["en", "zh-CN"],
+//     locales: ["en", "zh-CN", "zh-HK", "ja"],
 //     defaultLocale,
 //   });
 //   const response = handleI18nRouting(request);